psma/requirements.txt
amqp==5.0.6
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.0.5
Cerberus==1.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
debtcollector==2.2.0
fasteners==0.16
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
importlib-metadata==4.0.1
importlib-resources==5.1.2
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
netaddr==0.8.0
netifaces==0.10.9
oslo.concurrency==4.4.0
oslo.config==8.6.0
oslo.i18n==5.0.1
oslo.utils==4.8.0
packaging==20.9
paramiko==2.7.2
pbr==5.6.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyaml==20.4.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
six==1.15.0
SQLAlchemy==1.4.12
stevedore==3.3.0
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
virtualbox==2.1.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
docker==5.0.0